# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Market Intelligence API for TSR20 and EUR/USD
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AnalysisRecordPayload = { [key: string]: unknown };

export interface AnalysisRecord {
  id: number;
  market: string;
  run_mode: string;
  payload: AnalysisRecordPayload;
  created_at: string;
}

export interface AnalysisSummary {
  id: number;
  market: string;
  run_mode: string;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  conviction?: string | null;
  created_at: string;
}

export type WeeklyReportPayload = { [key: string]: unknown };

export interface WeeklyReport {
  id: number;
  market: string;
  week_start: string;
  payload: WeeklyReportPayload;
  created_at: string;
}

export type ScorecardRecordPayload = { [key: string]: unknown };

export interface ScorecardRecord {
  id: number;
  market: string;
  payload: ScorecardRecordPayload;
  created_at: string;
}

export interface FlashInput {
  headline: string;
}

export type FlashResultPayload = { [key: string]: unknown };

export interface FlashResult {
  id: number;
  headline_input: string;
  payload: FlashResultPayload;
  created_at: string;
}

export interface TriggerResponse {
  message: string;
  market: string;
}

export interface OutcomeInput {
  market: string;
  analysis_id: number;
  outcome_text: string;
}

export interface OutcomeRecord {
  id: number;
  market: string;
  analysis_id: number;
  outcome_text: string;
  entered_at: string;
}

export interface AnalysisWithOutcome {
  id: number;
  market: string;
  run_mode: string;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  conviction?: string | null;
  created_at: string;
  /** @nullable */
  outcome_id?: number | null;
  /** @nullable */
  outcome_text?: string | null;
  /** @nullable */
  outcome_entered_at?: string | null;
}

export interface SystemStatus {
  api_key_configured: boolean;
  scheduler_running: boolean;
  markets: string[];
  refresh_hours: number;
  model: string;
}