# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Market Intelligence API for TSR20 and EUR/USD
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analyses
    description: Market analyses
  - name: flash
    description: Flash alerts
  - name: weekly
    description: Weekly reports
  - name: scorecard
    description: Scorecards
  - name: outcomes
    description: Outcomes tracking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /latest/{market}:
    get:
      operationId: getLatestAnalysis
      tags: [analyses]
      summary: Get the most recent analysis for a market
      parameters:
        - name: market
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRecord"
        "404":
          description: No analysis found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /history/{market}:
    get:
      operationId: getAnalysisHistory
      tags: [analyses]
      summary: List past analyses for a market (returns last 20)
      parameters:
        - name: market
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisSummary"

  /analysis/{id}:
    get:
      operationId: getAnalysisById
      tags: [analyses]
      summary: Get full analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /weekly/{market}:
    get:
      operationId: getWeeklyReport
      tags: [weekly]
      summary: Get the latest weekly report for a market
      parameters:
        - name: market
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /scorecard/{market}:
    get:
      operationId: getScorecard
      tags: [scorecard]
      summary: Get the latest scorecard for a market
      parameters:
        - name: market
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScorecardRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /flash:
    post:
      operationId: runFlashAlert
      tags: [flash]
      summary: Run a flash assessment on a breaking headline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashInput"
      responses:
        "200":
          description: Flash result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashResult"
        "503":
          description: API not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /run-now/{market}:
    post:
      operationId: triggerMarketRun
      tags: [analyses]
      summary: Manually trigger a master analysis run for a market
      parameters:
        - name: market
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Run triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerResponse"

  /outcomes:
    post:
      operationId: recordOutcome
      tags: [outcomes]
      summary: Record actual outcome for a past analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutcomeInput"
      responses:
        "201":
          description: Outcome recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutcomeRecord"

  /scorecard/run/{market}:
    post:
      operationId: runScorecard
      tags: [scorecard]
      summary: Generate a scorecard for a market using recorded outcomes
      parameters:
        - name: market
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scorecard generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScorecardRecord"
        "400":
          description: Not enough data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyses-with-outcomes/{market}:
    get:
      operationId: getAnalysesWithOutcomes
      tags: [outcomes]
      summary: Get analyses along with their recorded outcomes for a market
      parameters:
        - name: market
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of analyses with optional outcomes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisWithOutcome"

  /status:
    get:
      operationId: getSystemStatus
      tags: [health]
      summary: Get system status including scheduler and API key
      responses:
        "200":
          description: System status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatus"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AnalysisRecord:
      type: object
      properties:
        id:
          type: integer
        market:
          type: string
        run_mode:
          type: string
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
      required:
        - id
        - market
        - run_mode
        - payload
        - created_at

    AnalysisSummary:
      type: object
      properties:
        id:
          type: integer
        market:
          type: string
        run_mode:
          type: string
        headline:
          type: ["string", "null"]
        conviction:
          type: ["string", "null"]
        created_at:
          type: string
      required:
        - id
        - market
        - run_mode
        - created_at

    WeeklyReport:
      type: object
      properties:
        id:
          type: integer
        market:
          type: string
        week_start:
          type: string
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
      required:
        - id
        - market
        - week_start
        - payload
        - created_at

    ScorecardRecord:
      type: object
      properties:
        id:
          type: integer
        market:
          type: string
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
      required:
        - id
        - market
        - payload
        - created_at

    FlashInput:
      type: object
      properties:
        headline:
          type: string
      required:
        - headline

    FlashResult:
      type: object
      properties:
        id:
          type: integer
        headline_input:
          type: string
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
      required:
        - id
        - headline_input
        - payload
        - created_at

    TriggerResponse:
      type: object
      properties:
        message:
          type: string
        market:
          type: string
      required:
        - message
        - market

    OutcomeInput:
      type: object
      properties:
        market:
          type: string
        analysis_id:
          type: integer
        outcome_text:
          type: string
      required:
        - market
        - analysis_id
        - outcome_text

    OutcomeRecord:
      type: object
      properties:
        id:
          type: integer
        market:
          type: string
        analysis_id:
          type: integer
        outcome_text:
          type: string
        entered_at:
          type: string
      required:
        - id
        - market
        - analysis_id
        - outcome_text
        - entered_at

    AnalysisWithOutcome:
      type: object
      properties:
        id:
          type: integer
        market:
          type: string
        run_mode:
          type: string
        headline:
          type: ["string", "null"]
        conviction:
          type: ["string", "null"]
        created_at:
          type: string
        outcome_id:
          type: ["integer", "null"]
        outcome_text:
          type: ["string", "null"]
        outcome_entered_at:
          type: ["string", "null"]
      required:
        - id
        - market
        - run_mode
        - created_at

    SystemStatus:
      type: object
      properties:
        api_key_configured:
          type: boolean
        scheduler_running:
          type: boolean
        markets:
          type: array
          items:
            type: string
        refresh_hours:
          type: integer
        model:
          type: string
      required:
        - api_key_configured
        - scheduler_running
        - markets
        - refresh_hours
        - model