# lib/api-zod/src/generated/types/analysisRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Market Intelligence API for TSR20 and EUR/USD
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisRecordPayload } from './analysisRecordPayload';

export interface AnalysisRecord {
  id: number;
  market: string;
  run_mode: string;
  payload: AnalysisRecordPayload;
  created_at: string;
}