# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Market Intelligence API for TSR20 and EUR/USD
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSummary {
  id: number;
  market: string;
  run_mode: string;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  conviction?: string | null;
  created_at: string;
}