# lib/api-zod/src/generated/types/analysisWithOutcome.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Market Intelligence API for TSR20 and EUR/USD
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisWithOutcome {
  id: number;
  market: string;
  run_mode: string;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  conviction?: string | null;
  created_at: string;
  /** @nullable */
  outcome_id?: number | null;
  /** @nullable */
  outcome_text?: string | null;
  /** @nullable */
  outcome_entered_at?: string | null;
}