# lib/api-zod/src/generated/types/flashResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Market Intelligence API for TSR20 and EUR/USD
 * OpenAPI spec version: 0.1.0
 */
import type { FlashResultPayload } from './flashResultPayload';

export interface FlashResult {
  id: number;
  headline_input: string;
  payload: FlashResultPayload;
  created_at: string;
}