# lib/api-zod/src/generated/types/scorecardRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Market Intelligence API for TSR20 and EUR/USD
 * OpenAPI spec version: 0.1.0
 */
import type { ScorecardRecordPayload } from './scorecardRecordPayload';

export interface ScorecardRecord {
  id: number;
  market: string;
  payload: ScorecardRecordPayload;
  created_at: string;
}