# lib/api-zod/src/generated/types/weeklyReport.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Market Intelligence API for TSR20 and EUR/USD
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklyReportPayload } from './weeklyReportPayload';

export interface WeeklyReport {
  id: number;
  market: string;
  week_start: string;
  payload: WeeklyReportPayload;
  created_at: string;
}